Created CorallerConfig and fixed errors with Elavator, flicker, and angler cause of it
